@@ -6,13 +6,13 @@ version: 4.0.1+4.0.1
 homepage: http://placar.iabetsoficial.com.br/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"  
+  sdk: ">=2.17.5 <3.0.0"  
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^0.23.0+1
-  mask_text_input_formatter: ^2.0.0
+  flutter_svg: ^1.1.1
+  mask_text_input_formatter: ^2.3.0
   image_picker: 
   path_provider: ^2.0.9
   flutter_spinkit: ^5.1.0
@@ -22,8 +22,8 @@ dependencies:
   dio: 4.0.3
 
   #state management
-  flutter_bloc: ^6.0.5
-  equatable: ^1.2.5
+  flutter_bloc: ^8.0.1
+  equatable: ^2.0.3
 
   #firebase
   firebase_auth: 
@@ -54,15 +54,17 @@ dependencies:
   file_picker: ^4.4.0
   timeago: ^3.2.2
 
+  #splash
+  flutter_native_splash: ^2.2.3+1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
- 
-  flutter_native_splash: ^1.2.0
 
 flutter_native_splash:
   background_image: assets/images/splash.png
   android: true
+  ios: true
 
 flutter:
   uses-material-design: true    
