@@ -3,9 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:ia_bet/domain/entities/double_config.dart';
 
 class SettingsController {
-  final formkey = GlobalKey<FormState>();
+  final GlobalKey<FormState> formkey = GlobalKey<FormState>();
+  final GlobalKey<FormState> formkeyEdit = GlobalKey<FormState>();
   final ValueNotifier<bool> stopGainIsOn = ValueNotifier<bool>(true);
   final ValueNotifier<bool> stopLossIsOn = ValueNotifier<bool>(true);
+  final ValueNotifier<bool> elevationIsOn = ValueNotifier<bool>(true);
+  final ValueNotifier<bool> galesIsOn = ValueNotifier<bool>(true);
   final TextEditingController stopGainController = TextEditingController();
   final TextEditingController stoplossController = TextEditingController();
   final TextEditingController firstBetPriceController = TextEditingController();
@@ -14,7 +17,13 @@ class SettingsController {
       TextEditingController();
   final TextEditingController newFirstBetWhiteController =
       TextEditingController();
+  final TextEditingController editFirstBetPriceController =
+      TextEditingController();
+  final TextEditingController editFirstBetWhiteController =
+      TextEditingController();
   final TextEditingController multiplierController = TextEditingController();
+  final TextEditingController editMultiplierController =
+      TextEditingController();
 
   Future<void> init(DoubleConfigEntity doubleConfig) async {
     stopGainController.text = doubleConfig.amountStopGain.toString();
@@ -23,9 +32,13 @@ class SettingsController {
     firstBetWhiteController.text = doubleConfig.entryWhiteAmount.toString();
     stopGainIsOn.value = doubleConfig.isActiveStopGain;
     stopLossIsOn.value = doubleConfig.isActiveStopLoss;
+    elevationIsOn.value = doubleConfig.isActiveElevation;
+    galesIsOn.value = doubleConfig.isActiveGale;
   }
 
   void dispose() {
+    galesIsOn.dispose();
+    elevationIsOn.dispose();
     stopGainIsOn.dispose();
     stopLossIsOn.dispose();
     stopGainController.dispose();
@@ -35,5 +48,8 @@ class SettingsController {
     newFirstBetPriceController.dispose();
     newFirstBetWhiteController.dispose();
     multiplierController.dispose();
+    editFirstBetPriceController.dispose();
+    editFirstBetWhiteController.dispose();
+    editMultiplierController.dispose();
   }
 }
