@@ -18,8 +18,22 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   Future<List<Map<String, dynamic>>> makeButtons() async => [
-        {'title': 'canais', 'route': CanaisPage(userInfo: await getUser())},
-        {'title': 'blaze', 'route': BlazePage()},
+        {
+          'title': 'Canais',
+          'icon': SvgPicture.asset(
+            'assets/images/chat.svg',
+            width: 30,
+          ),
+          'route': CanaisPage(userInfo: await getUser())
+        },
+        {
+          'title': 'Blaze',
+          'icon': SvgPicture.asset(
+            'assets/images/blaze.svg',
+            width: 30,
+          ),
+          'route': BlazePage(),
+        },
       ];
 
   Future<UserEntity> getUser() async =>
@@ -29,51 +43,91 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     final Size size = MediaQuery.of(context).size;
 
-    return Scaffold(
-      appBar: AppBar(
-        title: Row(
-          children: [
-            SvgPicture.asset(
-              'assets/images/logo.svg',
-              width: 100,
-            ),
-            Text('Bets')
-          ],
-        ),
-      ),
-      body: FutureBuilder(
-        future: makeButtons(),
-        builder: (BuildContext context,
-            AsyncSnapshot<List<Map<String, dynamic>>> snapshot) {
-          if (snapshot.connectionState == ConnectionState.done &&
-              snapshot.hasData) {
-            return ListView.builder(
-              itemCount: snapshot.data!.length,
-              itemBuilder: (BuildContext context, int index) => Padding(
-                padding: EdgeInsets.only(
-                    left: size.width * 0.2,
-                    right: size.width * 0.2,
-                    top: size.height * 0.05),
-                child: ElevatedButton(
-                  onPressed: () => Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (BuildContext context) =>
-                            snapshot.data![index]['route']),
+    return WillPopScope(
+        onWillPop: () async => false,
+        child: Scaffold(
+            appBar: AppBar(
+              title: Row(
+                children: [
+                  SvgPicture.asset(
+                    'assets/images/logo.svg',
+                    width: 100,
+                    color: Colors.white,
                   ),
-                  child: Center(
-                    child: Text(snapshot.data![index]['title']),
-                  ),
-                ),
+                ],
               ),
-            );
-          } else {
-            return Center(
-              child: CircularProgressIndicator(),
-            );
-          }
-        },
-      ),
-    );
+            ),
+            body: Stack(
+              children: [
+                Container(
+                    decoration: BoxDecoration(
+                  image: DecorationImage(
+                    image: AssetImage('assets/images/background_home_page.png'),
+                    fit: BoxFit.cover,
+                  ),
+                )),
+                FutureBuilder(
+                  future: makeButtons(),
+                  builder: (BuildContext context,
+                      AsyncSnapshot<List<Map<String, dynamic>>> snapshot) {
+                    if (snapshot.connectionState == ConnectionState.done &&
+                        snapshot.hasData) {
+                      return ListView.builder(
+                        itemCount: snapshot.data!.length,
+                        itemBuilder: (BuildContext context, int index) =>
+                            SizedBox(
+                          height: size.height * 0.09,
+                          child: ElevatedButton(
+                            style: ElevatedButton.styleFrom(
+                              elevation: 0,
+                              padding: EdgeInsets.zero,
+                              primary: Colors.grey.shade300.withAlpha(120),
+                            ),
+                            onPressed: () => Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                  builder: (BuildContext context) =>
+                                      snapshot.data![index]['route']),
+                            ),
+                            child: Padding(
+                              padding: EdgeInsets.symmetric(
+                                  horizontal: size.width * 0.07),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Flexible(
+                                      child: snapshot.data![index]['icon']),
+                                  Flexible(
+                                    child: Text(
+                                      snapshot.data![index]['title'],
+                                      style: TextStyle(
+                                          color: Colors.black, fontSize: 18),
+                                    ),
+                                  ),
+                                  const Spacer(
+                                    flex: 1,
+                                  ),
+                                  Flexible(
+                                    child: Icon(
+                                      Icons.arrow_forward_ios_rounded,
+                                      color: Colors.black,
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                      );
+                    } else {
+                      return Center(
+                        child: CircularProgressIndicator(),
+                      );
+                    }
+                  },
+                ),
+              ],
+            )));
   }
 }
